renamed gui_components to gui for convenience
