+++ src/gui/__init__.py
@@ -1,0 +1,9 @@
+import src.gui.menu.main as menu
+import src.gui.view.main as view
+import src.gui.edit.main as edit
+import src.gui.settings.main as settings
+
+Menu = menu.Menu
+View = view.View
+Edit = edit.Edit
+Settings = settings.Settings
