+++ src/gui/settings/keybindings.py
@@ -1,6 +1,6 @@
 import tkinter as tk
 import keyboard as kb
-from src.gui_components.interfaces import LabelFrame, Frame
+from src.gui.interfaces import LabelFrame, Frame
 from src.common import utils
 from src.common.interfaces import Configurable
